ADDED: 1 to 5
There is now an inventory shown on the screen using the "I" key and you can equip the different items.

+++ Assets/Game/Scripts/Entity/Entity.cs
@@ -7,7 +7,7 @@ public abstract class Entity : MonoBehaviour, IDamageable
 
     protected float _currentHealth;
 
-    public virtual void Heal(int amount)
+    public virtual void Heal(float amount)
     {
         _currentHealth += amount;
         if(_currentHealth > _initHealth)
@@ -29,7 +29,7 @@ public abstract class Entity : MonoBehaviour, IDamageable
 
     }
 
-    public abstract void TakeDamage(int amount);
+    public abstract void TakeDamage(float amount);
 
     public virtual void Shoot() { }
 }
